+++ src/CryptoNoteCore/BlockchainStorage.cpp
@@ -41,17 +41,17 @@ void BlockchainStorage::pushBlock(RawBlock&& rawBlock) {
   internalStorage->pushBlock(std::move(rawBlock));
 }
 
-RawBlock BlockchainStorage::getBlockByIndex(uint32_t index) const {
+RawBlock BlockchainStorage::getBlockByIndex(uint64_t index) const {
   return internalStorage->getBlockByIndex(index);
 }
 
-uint32_t BlockchainStorage::getBlockCount() const {
+uint64_t BlockchainStorage::getBlockCount() const {
   return internalStorage->getBlockCount();
 }
 
 //Returns MemoryBlockchainStorage with elements from [splitIndex, blocks.size() - 1].
 //Original MemoryBlockchainStorage will contain elements from [0, splitIndex - 1].
-std::unique_ptr<BlockchainStorage> BlockchainStorage::splitStorage(uint32_t splitIndex) {
+std::unique_ptr<BlockchainStorage> BlockchainStorage::splitStorage(uint64_t splitIndex) {
   std::unique_ptr<BlockchainStorage> newStorage(new BlockchainStorage(internalStorage->splitStorage(splitIndex)));
   return newStorage;
 }
